+++ src/aks-preview/azext_aks_preview/tests/latest/test_aks_commands.py
@@ -17,6 +17,11 @@ from azext_aks_preview.tests.latest.custom_preparers import (
 )
 from azext_aks_preview.tests.latest.recording_processors import KeyReplacer
 from azure.cli.command_modules.acs._format import version_to_tuple
+from azure.cli.command_modules.acs._helpers import (
+    get_shared_control_plane_identity,
+    get_shared_kubelet_identity,
+    use_shared_identity,
+)
 from azure.cli.core.azclierror import ClientRequestError
 from azure.core.exceptions import (HttpResponseError)
 from azure.cli.testsdk import CliTestError, ScenarioTest, live_only
@@ -37,17 +42,14 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
     def _get_versions(self, location):
         """Return the previous and current Kubernetes minor release versions, such as ("1.11.6", "1.12.4")."""
-        versions = self.cmd(
-            "az aks get-versions -l {} --query 'values[*].patchVersions.keys(@)[]'".format(
-                location
-            )
+        supported_versions = self.cmd(
+            "az aks get-versions -l {} --query 'values[*].patchVersions.keys(@)[]'".format(location)
         ).get_output_in_json()
-        # sort by semantic version, from newest to oldest
-        versions = sorted(versions, key=version_to_tuple, reverse=True)
-        upgrade_version = versions[0]
+        sorted_supported_versions = sorted(supported_versions, key=version_to_tuple, reverse=True)
+        upgrade_version = sorted_supported_versions[0]
         # find the first version that doesn't start with the latest major.minor.
-        prefix = upgrade_version[: upgrade_version.rfind(".")]
-        create_version = next(x for x in versions if not x.startswith(prefix))
+        prefix = upgrade_version[:upgrade_version.rfind('.')]
+        create_version = next(x for x in sorted_supported_versions if not x.startswith(prefix))
         return create_version, upgrade_version
 
     def _get_version_in_range(
@@ -85,6 +87,69 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         # sort by semantic version, from newest to oldest
         lts_versions = sorted(lts_versions, key=lambda x: list(map(int, x.split("."))), reverse=True)
         return lts_versions[0] if lts_versions else None
+
+    def _get_user_assigned_identity(
+        self,
+        resource_group,
+        designated_identity=None,
+        excluded_identity=None,
+        use_for_kubelet=False,
+        paired_control_plane_identity=None,
+        kubelet_identity_index_shift=0,
+    ):
+        # If designated_identity is provided, it will be used.
+        if designated_identity:
+            return designated_identity
+        # If the environment variable USE_SHARED_IDENTITY is set, the shared identity will be used.
+        if use_shared_identity():
+            return (
+                get_shared_kubelet_identity(
+                    excluded_identity=excluded_identity,
+                    shared_control_plane_identity=paired_control_plane_identity,
+                    index_shift=kubelet_identity_index_shift,
+                )
+                if use_for_kubelet
+                else get_shared_control_plane_identity(
+                    excluded_identity=excluded_identity,
+                )
+            )
+        # Otherwise, a new user-assigned identity will be created.
+        identity_name = self.create_random_name("cli", 16)
+        identity = self.cmd(
+            "az identity create -g {} -n {}".format(resource_group, identity_name)
+        ).get_output_in_json()
+        return identity.get("id")
+
+    def _get_principal_id_of_user_assigned_identity(self, identity_id):
+        return self.cmd(
+            "identity show --ids {}".format(identity_id)
+        ).get_output_in_json()["principalId"]
+
+    def _get_client_id_of_user_assigned_identity(self, identity_id):
+        return self.cmd(
+            "identity show --ids {}".format(identity_id)
+        ).get_output_in_json()["clientId"]
+
+    def _get_test_identity_resource_id(self) -> str:
+        return os.getenv("IDENTITY_RESOURCE_ID")
+
+    def _get_test_identity_object_id(self) -> str:
+        test_identity_object_id = os.getenv("AZURE_CLI_LIVE_TEST_IDENTITY_OBJECT_ID")
+        if test_identity_object_id:
+            return os.getenv("AZURE_CLI_LIVE_TEST_IDENTITY_OBJECT_ID")
+        identity_resource_id = self._get_test_identity_resource_id()
+        # test executed in CI mode
+        if identity_resource_id:
+            test_identity_object_id = self._get_principal_id_of_user_assigned_identity(
+                identity_resource_id
+            )
+        # test executed in local mode
+        else:
+            test_identity_object_id = self.cmd(
+                "az ad signed-in-user show"
+            ).get_output_in_json()["id"]
+        os.environ["AZURE_CLI_LIVE_TEST_IDENTITY_OBJECT_ID"] = test_identity_object_id
+        return test_identity_object_id
 
     def _sort_revisions(self, revisions):
         def _convert_revision_to_semver(rev):
@@ -2848,7 +2913,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_nodepool_add_with_disable_windows_outbound_nat(
         self, resource_group, resource_group_location
@@ -3753,7 +3817,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_snapshot(self, resource_group, resource_group_location):
         print(resource_group_location)
@@ -3870,7 +3933,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_snapshot_upgrade(self, resource_group, resource_group_location):
         print(resource_group_location)
@@ -3900,6 +3962,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             "--nodepool-name {nodepool_name} "
             "--node-count 1 "
             "-k {upgrade_k8s_version} "
+            "--load-balancer-sku=standard --vm-set-type=virtualmachinescalesets --network-plugin=azure "
             "--ssh-key-value={ssh_key_value} -o json"
         )
         response = self.cmd(
@@ -3961,6 +4024,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             "--nodepool-name {nodepool_name} "
             "--node-count 1 "
             "-k {k8s_version} "
+            "--load-balancer-sku=standard --vm-set-type=virtualmachinescalesets --network-plugin=azure "
             "--ssh-key-value={ssh_key_value} -o json"
         )
         self.cmd(
@@ -4000,7 +4064,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_snapshot_update(self, resource_group, resource_group_location):
         print(resource_group_location)
@@ -4128,7 +4191,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_skip_gpu_driver_install(self, resource_group, resource_group_location):
         print(resource_group_location)
@@ -4864,7 +4926,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             "aks update --resource-group={resource_group} --name={name} --enable-managed-identity --yes",
             checks=[
                 self.check("provisioningState", "Succeeded"),
-                self.check("identity.type", "SystemAssigned"),
+                self.check("identity.type", "UserAssigned" if use_shared_identity() else "SystemAssigned"),
             ],
         )
 
@@ -5020,7 +5082,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             "aks update --resource-group={resource_group} --name={name} --enable-managed-identity --yes",
             checks=[
                 self.check("provisioningState", "Succeeded"),
-                self.check("identity.type", "SystemAssigned"),
+                self.check("identity.type", "UserAssigned" if use_shared_identity() else "SystemAssigned"),
             ],
         )
 
@@ -5142,9 +5204,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         if user_assigned_identity:
-            uai_cmd = f"identity create -g {resource_group} -n {aks_name}_uai"
-            resp = self.cmd(uai_cmd).get_output_in_json()
-            identity_id = resp["id"]
+            identity_id = self._get_user_assigned_identity(resource_group)
             print("********************")
             print(f"identity_id: {identity_id}")
             print("********************")
@@ -5383,9 +5443,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         if user_assigned_identity:
-            uai_cmd = f"identity create -g {resource_group} -n {aks_name}_uai"
-            resp = self.cmd(uai_cmd).get_output_in_json()
-            identity_id = resp["id"]
+            identity_id = self._get_user_assigned_identity(resource_group)
             print("********************")
             print(f"identity_id: {identity_id}")
             print("********************")
@@ -6045,14 +6103,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create identity
-        create_identity = (
-            "identity create --resource-group={resource_group} --name={identity_name}"
-        )
-        identity = self.cmd(
-            create_identity, checks=[self.check("name", identity_name)]
-        ).get_output_in_json()
-        identity_id = identity["principalId"]
-        identity_resource_id = identity["id"]
+        identity_resource_id = self._get_user_assigned_identity(resource_group)
+        identity_id = self._get_principal_id_of_user_assigned_identity(identity_resource_id)
         assert identity_id is not None
         self.kwargs.update(
             {
@@ -6404,17 +6456,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             }
         )
 
-        # create identity
-        cmd = "identity create --resource-group={resource_group} --name={identity_name} --location={location}"
-        application_identity = self.cmd(
-            cmd, checks=[self.check("name", identity_name)]
-        ).get_output_in_json()
-        self.kwargs.update(
-            {
-                "application_identity_id": application_identity["id"],
-            }
-        )
-
         # create
         create_cmd = (
             "aks create --resource-group={resource_group} --name={name} --location={location} "
@@ -6424,12 +6465,25 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             "--ssh-key-value={ssh_key_value} "
             "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/EnablePodIdentityPreview"
         )
-        self.cmd(
+        mc = self.cmd(
             create_cmd,
             checks=[
                 self.check("provisioningState", "Succeeded"),
                 self.check("podIdentityProfile.enabled", True),
             ],
+        ).get_output_in_json()
+        control_plane_identity = None
+        if use_shared_identity():
+            control_plane_identity = list(mc["identity"]["userAssignedIdentities"].keys())[0]
+
+        # create identity
+        application_identity_id = self._get_user_assigned_identity(resource_group, use_for_kubelet=True, paired_control_plane_identity=control_plane_identity)
+        application_identity_client_id = self._get_client_id_of_user_assigned_identity(application_identity_id)
+        application_identity_principal_id = self._get_principal_id_of_user_assigned_identity(application_identity_id)
+        self.kwargs.update(
+            {
+                "application_identity_id": application_identity_id,
+            }
         )
 
         # pod identity: add
@@ -6456,15 +6510,16 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
                 ),
                 self.check(
                     "podIdentityProfile.userAssignedIdentities[0].identity.clientId",
-                    application_identity["clientId"],
+                    application_identity_client_id,
                 ),
                 self.check(
                     "podIdentityProfile.userAssignedIdentities[0].identity.objectId",
-                    application_identity["principalId"],
+                    application_identity_principal_id,
                 ),
                 self.check(
                     "podIdentityProfile.userAssignedIdentities[0].identity.resourceId",
-                    application_identity["id"],
+                    application_identity_id,
+                    case_sensitive=False,
                 ),
             ],
         )
@@ -6514,15 +6569,16 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
                 ),
                 self.check(
                     "podIdentityProfile.userAssignedIdentities[0].identity.clientId",
-                    application_identity["clientId"],
+                    application_identity_client_id,
                 ),
                 self.check(
                     "podIdentityProfile.userAssignedIdentities[0].identity.objectId",
-                    application_identity["principalId"],
+                    application_identity_principal_id,
                 ),
                 self.check(
                     "podIdentityProfile.userAssignedIdentities[0].identity.resourceId",
-                    application_identity["id"],
+                    application_identity_id,
+                    case_sensitive=False,
                 ),
             ],
         )
@@ -6624,12 +6680,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create control plane identity
-        control_plane_identity = "identity create --resource-group={resource_group} --name={control_plane_identity_name}"
-        c_identity = self.cmd(
-            control_plane_identity,
-            checks=[self.check("name", control_plane_identity_name)],
-        ).get_output_in_json()
-        control_plane_identity_resource_id = c_identity["id"]
+        control_plane_identity_resource_id = self._get_user_assigned_identity(resource_group)
         assert control_plane_identity_resource_id is not None
         self.kwargs.update(
             {
@@ -6638,11 +6689,11 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create kubelet identity
-        kubelet_identity = "identity create --resource-group={resource_group} --name={kubelet_identity_name}"
-        k_identity = self.cmd(
-            kubelet_identity, checks=[self.check("name", kubelet_identity_name)]
-        ).get_output_in_json()
-        kubelet_identity_resource_id = k_identity["id"]
+        kubelet_identity_resource_id = self._get_user_assigned_identity(
+            resource_group,
+            use_for_kubelet=True,
+            paired_control_plane_identity=control_plane_identity_resource_id,
+        )
         assert kubelet_identity_resource_id is not None
         self.kwargs.update(
             {
@@ -6671,11 +6722,12 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create new kubelet identity
-        new_kubelet_identity = "identity create --resource-group={resource_group} --name={new_kubelet_identity_name}"
-        new_identity = self.cmd(
-            new_kubelet_identity, checks=[self.check("name", new_kubelet_identity_name)]
-        ).get_output_in_json()
-        new_kubelet_identity_resource_id = new_identity["id"]
+        new_kubelet_identity_resource_id = self._get_user_assigned_identity(
+            resource_group,
+            use_for_kubelet=True,
+            paired_control_plane_identity=control_plane_identity_resource_id,
+            kubelet_identity_index_shift=1,
+        )
         assert new_kubelet_identity_resource_id is not None
         self.kwargs.update(
             {
@@ -6749,7 +6801,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
     @AKSCustomResourceGroupPreparer(
         random_name_length=17,
         name_prefix="clitest",
-        location="eastus",
+        location="uksouth",
         preserve_default_location=True,
     )
     def test_aks_enable_utlra_ssd(self, resource_group, resource_group_location):
@@ -7063,7 +7115,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_with_overlay_network_plugin_mode(
         self, resource_group, resource_group_location
@@ -7109,7 +7160,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_create_with_network_dataplane_cilium(
         self, resource_group, resource_group_location
@@ -7159,7 +7209,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_with_enable_cilium_dataplane(
         self, resource_group, resource_group_location
@@ -7207,7 +7256,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_dualstack_with_default_network(
         self, resource_group, resource_group_location
@@ -7412,7 +7460,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_and_update_ipv6_count(
         self, resource_group, resource_group_location
@@ -7504,7 +7551,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_azure_cni_overlay_migration(
         self, resource_group, resource_group_location
@@ -7567,7 +7613,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_azure_cni_overlay_migration_from_kubenet(
         self, resource_group, resource_group_location
@@ -7629,7 +7674,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_migrate_cluster_to_cilium_dataplane(
         self, resource_group, resource_group_location
@@ -7688,7 +7732,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_or_update_with_load_balancer_backend_pool_type(
         self, resource_group, resource_group_location
@@ -7741,7 +7784,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_or_update_with_health_probe_mode(
         self, resource_group, resource_group_location
@@ -8865,10 +8907,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create user-assigned identity
-        create_identity = "identity create --resource-group={resource_group} --name={identity_name} -o json"
-        identity = self.cmd(create_identity).get_output_in_json()
-        identity_id = identity["id"]
-        identity_object_id = identity["principalId"]
+        identity_id = self._get_user_assigned_identity(resource_group)
+        identity_object_id = self._get_principal_id_of_user_assigned_identity(identity_id)
         assert identity_id is not None
         assert identity_object_id is not None
         self.kwargs.update(
@@ -8880,12 +8920,20 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         # create key vault and key
         create_keyvault = (
-            "keyvault create --resource-group={resource_group} --name={kv_name} -o json"
+            "keyvault create --resource-group={resource_group} --name={kv_name} --enable-rbac-authorization=false --no-self-perms -o json"
         )
         self.cmd(
             create_keyvault,
             checks=[self.check("properties.provisioningState", "Succeeded")],
         )
+
+        # set access policy for test identity
+        test_identity_object_id = self._get_test_identity_object_id()
+        test_identity_access_policy = 'keyvault set-policy --resource-group={resource_group} --name={kv_name} ' \
+                                      '--key-permissions all --object-id ' + test_identity_object_id
+        self.cmd(test_identity_access_policy, checks=[
+            self.check('properties.provisioningState', 'Succeeded')
+        ])
 
         create_key = "keyvault key create -n kms --vault-name {kv_name} -o json"
         key = self.cmd(
@@ -8988,10 +9036,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create user-assigned identity
-        create_identity = "identity create --resource-group={resource_group} --name={identity_name} -o json"
-        identity = self.cmd(create_identity).get_output_in_json()
-        identity_id = identity["id"]
-        identity_object_id = identity["principalId"]
+        identity_id = self._get_user_assigned_identity(resource_group)
+        identity_object_id = self._get_principal_id_of_user_assigned_identity(identity_id)
         assert identity_id is not None
         assert identity_object_id is not None
         self.kwargs.update(
@@ -9003,12 +9049,20 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         # create key vault and key
         create_keyvault = (
-            "keyvault create --resource-group={resource_group} --name={kv_name} -o json"
+            "keyvault create --resource-group={resource_group} --name={kv_name} --enable-rbac-authorization=false --no-self-perms -o json"
         )
         self.cmd(
             create_keyvault,
             checks=[self.check("properties.provisioningState", "Succeeded")],
         )
+
+        # set access policy for test identity
+        test_identity_object_id = self._get_test_identity_object_id()
+        test_identity_access_policy = 'keyvault set-policy --resource-group={resource_group} --name={kv_name} ' \
+                                      '--key-permissions all --object-id ' + test_identity_object_id
+        self.cmd(test_identity_access_policy, checks=[
+            self.check('properties.provisioningState', 'Succeeded')
+        ])
 
         create_key = "keyvault key create -n kms --vault-name {kv_name} -o json"
         key = self.cmd(
@@ -9078,7 +9132,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_with_azurekeyvaultkms_private_key_vault(
         self, resource_group, resource_group_location
@@ -9097,10 +9150,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create user-assigned identity
-        create_identity = "identity create --resource-group={resource_group} --name={identity_name} -o json"
-        identity = self.cmd(create_identity).get_output_in_json()
-        identity_id = identity["id"]
-        identity_object_id = identity["principalId"]
+        identity_id = self._get_user_assigned_identity(resource_group)
+        identity_object_id = self._get_principal_id_of_user_assigned_identity(identity_id)
         assert identity_id is not None
         assert identity_object_id is not None
         self.kwargs.update(
@@ -9112,7 +9163,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         # create key vault and key
         create_keyvault = (
-            "keyvault create --resource-group={resource_group} --name={kv_name} -o json"
+            "keyvault create --resource-group={resource_group} --name={kv_name} --enable-rbac-authorization=false --no-self-perms -o json"
         )
         kv = self.cmd(
             create_keyvault,
@@ -9125,6 +9176,14 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
                 "kv_resource_id": kv_resource_id,
             }
         )
+
+        # set access policy for test identity
+        test_identity_object_id = self._get_test_identity_object_id()
+        test_identity_access_policy = 'keyvault set-policy --resource-group={resource_group} --name={kv_name} ' \
+                                      '--key-permissions all --object-id ' + test_identity_object_id
+        self.cmd(test_identity_access_policy, checks=[
+            self.check('properties.provisioningState', 'Succeeded')
+        ])
 
         create_key = "keyvault key create -n kms --vault-name {kv_name} -o json"
         key = self.cmd(
@@ -9250,7 +9309,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_update_with_azurekeyvaultkms_private_key_vault(
         self, resource_group, resource_group_location
@@ -9269,10 +9327,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create user-assigned identity
-        create_identity = "identity create --resource-group={resource_group} --name={identity_name} -o json"
-        identity = self.cmd(create_identity).get_output_in_json()
-        identity_id = identity["id"]
-        identity_object_id = identity["principalId"]
+        identity_id = self._get_user_assigned_identity(resource_group)
+        identity_object_id = self._get_principal_id_of_user_assigned_identity(identity_id)
         assert identity_id is not None
         assert identity_object_id is not None
         self.kwargs.update(
@@ -9284,7 +9340,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         # create key vault and key
         create_keyvault = (
-            "keyvault create --resource-group={resource_group} --name={kv_name} -o json"
+            "keyvault create --resource-group={resource_group} --name={kv_name} --enable-rbac-authorization=false --no-self-perms -o json"
         )
         kv = self.cmd(
             create_keyvault,
@@ -9297,6 +9353,14 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
                 "kv_resource_id": kv_resource_id,
             }
         )
+
+        # set access policy for test identity
+        test_identity_object_id = self._get_test_identity_object_id()
+        test_identity_access_policy = 'keyvault set-policy --resource-group={resource_group} --name={kv_name} ' \
+                                      '--key-permissions all --object-id ' + test_identity_object_id
+        self.cmd(test_identity_access_policy, checks=[
+            self.check('properties.provisioningState', 'Succeeded')
+        ])
 
         create_key = "keyvault key create -n kms --vault-name {kv_name} -o json"
         key = self.cmd(
@@ -9386,7 +9450,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_with_azurekeyvaultkms_private_cluster_v1_private_key_vault(
         self, resource_group, resource_group_location
@@ -9405,10 +9468,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create user-assigned identity
-        create_identity = "identity create --resource-group={resource_group} --name={identity_name} -o json"
-        identity = self.cmd(create_identity).get_output_in_json()
-        identity_id = identity["id"]
-        identity_object_id = identity["principalId"]
+        identity_id = self._get_user_assigned_identity(resource_group)
+        identity_object_id = self._get_principal_id_of_user_assigned_identity(identity_id)
         assert identity_id is not None
         assert identity_object_id is not None
         self.kwargs.update(
@@ -9420,7 +9481,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         # create key vault and key
         create_keyvault = (
-            "keyvault create --resource-group={resource_group} --name={kv_name} -o json"
+            "keyvault create --resource-group={resource_group} --name={kv_name} --enable-rbac-authorization=false --no-self-perms -o json"
         )
         kv = self.cmd(
             create_keyvault,
@@ -9433,6 +9494,14 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
                 "kv_resource_id": kv_resource_id,
             }
         )
+
+        # set access policy for test identity
+        test_identity_object_id = self._get_test_identity_object_id()
+        test_identity_access_policy = 'keyvault set-policy --resource-group={resource_group} --name={kv_name} ' \
+                                      '--key-permissions all --object-id ' + test_identity_object_id
+        self.cmd(test_identity_access_policy, checks=[
+            self.check('properties.provisioningState', 'Succeeded')
+        ])
 
         create_key = "keyvault key create -n kms --vault-name {kv_name} -o json"
         key = self.cmd(
@@ -9575,10 +9644,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # create user-assigned identity
-        create_identity = "identity create --resource-group={resource_group} --name={identity_name} -o json"
-        identity = self.cmd(create_identity).get_output_in_json()
-        identity_id = identity["id"]
-        identity_object_id = identity["principalId"]
+        identity_id = self._get_user_assigned_identity(resource_group)
+        identity_object_id = self._get_principal_id_of_user_assigned_identity(identity_id)
         assert identity_id is not None
         assert identity_object_id is not None
         self.kwargs.update(
@@ -9590,12 +9657,20 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         # create key vault and key
         create_keyvault = (
-            "keyvault create --resource-group={resource_group} --name={kv_name} -o json"
+            "keyvault create --resource-group={resource_group} --name={kv_name} --enable-rbac-authorization=false --no-self-perms -o json"
         )
         self.cmd(
             create_keyvault,
             checks=[self.check("properties.provisioningState", "Succeeded")],
         )
+
+        # set access policy for test identity
+        test_identity_object_id = self._get_test_identity_object_id()
+        test_identity_access_policy = 'keyvault set-policy --resource-group={resource_group} --name={kv_name} ' \
+                                      '--key-permissions all --object-id ' + test_identity_object_id
+        self.cmd(test_identity_access_policy, checks=[
+            self.check('properties.provisioningState', 'Succeeded')
+        ])
 
         create_key = "keyvault key create -n kms --vault-name {kv_name} -o json"
         key = self.cmd(
@@ -9665,7 +9740,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_create_and_update_with_csi_drivers_extensibility(
         self, resource_group, resource_group_location
@@ -9770,7 +9844,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_create_with_standard_csi_drivers(
         self, resource_group, resource_group_location
@@ -9828,7 +9901,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_with_standard_blob_csi_driver(
         self, resource_group, resource_group_location
@@ -9880,7 +9952,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_and_update_with_blob_csi_driver(
         self, resource_group, resource_group_location
@@ -9979,7 +10050,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_with_csi_driver_v2(
         self, resource_group, resource_group_location
@@ -10024,7 +10094,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_and_update_csi_driver_to_v2(
         self, resource_group, resource_group_location
@@ -10129,7 +10198,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_with_apiserver_vnet_integration_public(
         self, resource_group, resource_group_location
@@ -10616,14 +10684,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
         # Get the object ID of the service principal
-        show_acct_cmd = "account show"
-        authenticated_acct = self.cmd(show_acct_cmd).get_output_in_json()
-        sp_name = authenticated_acct["user"]["name"]
-        show_sp_cmd = f"ad sp show --id {sp_name}"
-        sp = self.cmd(
-            show_sp_cmd, checks=[self.check("appId", sp_name)]
-        ).get_output_in_json()
-        sp_oid = sp["id"]
+        sp_oid = self._get_test_identity_object_id()
         print(f"objectid of service principal is {sp_oid}")
 
         # Install kubectl (for setting up service principal permissions, and required by the 'kollect' command).
@@ -10747,7 +10808,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
     @AKSCustomResourceGroupPreparer(
         random_name_length=17,
         name_prefix="clitest",
-        location="eastus",
+        location="uksouth",
         preserve_default_location=True,
     )
     def test_aks_availability_zones(self, resource_group, resource_group_location):
@@ -10984,8 +11045,10 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
 
         create_cmd = (
-            "aks create --resource-group={resource_group} --name={name} --location={location} --ssh-key-value={ssh_key_value} --node-vm-size={node_vm_size} "
-            "--node-count 3 --enable-azure-monitor-app-monitoring --output=json"
+            "aks create --resource-group={resource_group} --name={name} --location={location} "
+            "--ssh-key-value={ssh_key_value} --node-vm-size={node_vm_size} --node-count 3 "
+            "--enable-azure-monitor-app-monitoring "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/AzureMonitorAppMonitoringPreview"
         )
         self.cmd(
             create_cmd,
@@ -11265,7 +11328,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         # update: enable-azure-monitor-app-monitoring
         update_cmd = (
-            'aks update --resource-group={resource_group} --name={name} --yes --output=json --enable-azure-monitor-app-monitoring'
+            "aks update --resource-group={resource_group} --name={name} --yes --output=json --enable-azure-monitor-app-monitoring "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/AzureMonitorAppMonitoringPreview"
         )
         self.cmd(update_cmd, checks=[
             self.check('provisioningState', 'Succeeded'),
@@ -11276,7 +11340,8 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
 
         # update: disable-azure-monitor-app-monitoring
         update_cmd = (
-            'aks update --resource-group={resource_group} --name={name} --yes --output=json --disable-azure-monitor-app-monitoring'
+            "aks update --resource-group={resource_group} --name={name} --yes --output=json --disable-azure-monitor-app-monitoring "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/AzureMonitorAppMonitoringPreview"
         )
 
         self.cmd(update_cmd, checks=[
@@ -12987,7 +13052,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_update_enable_advanced_network_observability(
         self, resource_group, resource_group_location
@@ -13045,7 +13109,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_create_with_enable_advanced_network_observability(
         self, resource_group, resource_group_location
@@ -13089,7 +13152,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_create_with_enable_cost_analysis(
         self, resource_group, resource_group_location
@@ -13133,7 +13195,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_update_enable_cost_analysis(
         self, resource_group, resource_group_location
@@ -13196,7 +13257,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_create_node_provisioning_mode_manual(
         self, resource_group, resource_group_location
@@ -13240,7 +13300,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="westcentralus",
-        preserve_default_location=True,
     )
     def test_aks_update_node_provisioning_mode_manual(
         self, resource_group, resource_group_location
@@ -13292,7 +13351,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_create_with_custom_headers(
         self, resource_group, resource_group_location
@@ -13403,7 +13461,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_approuting_enable_disable(
         self, resource_group, resource_group_location
@@ -13465,7 +13522,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_approuting_enable_with_keyvault_secrets_provider_addon_and_keyvault_id(
         self, resource_group, resource_group_location
@@ -13533,7 +13589,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_approuting_update(self, resource_group, resource_group_location):
         """This test case exercises updating app routing addon in an AKS cluster."""
@@ -13647,7 +13702,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_approuting_update_with_monitoring_addon_enabled(self, resource_group, resource_group_location):
         """This test case exercises updating app routing addon in an AKS cluster with monitoring addon enabled."""
@@ -13720,7 +13774,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_approuting_zone_add_delete_list(
         self, resource_group, resource_group_location
@@ -13843,7 +13896,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="eastus",
-        preserve_default_location=True,
     )
     def test_aks_approuting_zone_update(self, resource_group, resource_group_location):
         """This test case exercises updating zones to app routing addon in an AKS cluster."""
@@ -13977,8 +14029,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
     @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest',
-                                    location='westus2', preserve_default_location=True)
+    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='westus2')
     def test_aks_create_with_enable_ai_toolchain_operator(self, resource_group,
                                                           resource_group_location):
         # reset the count so in replay mode the random names will start with 0
@@ -14008,8 +14059,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             'aks delete -g {resource_group} -n {name} --yes --no-wait', checks=[self.is_empty()])
 
     @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest',
-                                    location='westus2', preserve_default_location=True)
+    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='westus2')
     def test_aks_update_with_enable_ai_toolchain_operator(self, resource_group, resource_group_location):
         # reset the count so in replay mode the random names will start with 0
         self.test_resources_count = 0
@@ -14109,7 +14159,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_create_with_pod_ip_allocation_mode_static_block(
         self, resource_group, resource_group_location
@@ -14204,8 +14253,7 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         )
 
     @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest',
-                                    location='westus2', preserve_default_location=True)
+    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='westus2')
     def test_aks_check_network(self, resource_group, resource_group_location):
         # reset the count so in replay mode the random names will start with 0
         self.test_resources_count = 0
@@ -14522,23 +14570,15 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         self.cmd(create_dns_record_cmd)
 
         # create identity
-        create_cluster_idenity_cmd = (
-            "identity create --name {cluster_identity_name} --resource-group {resource_group} -o json"
-        )
-        cluster_idenity = self.cmd(create_cluster_idenity_cmd).get_output_in_json()
-        cluster_identity_id = cluster_idenity["id"]
+        cluster_identity_id = self._get_user_assigned_identity(resource_group)
         assert cluster_identity_id is not None
         self.kwargs.update(
             {
                 "cluster_identity_id": cluster_identity_id,
             }
         )
-        create_kubelet_idenity_cmd = (
-            "identity create --name {kubelet_identity_name} --resource-group {resource_group} -o json"
-        )
-        kubelet_identity = self.cmd(create_kubelet_idenity_cmd).get_output_in_json()
-        kubelet_identity_id = kubelet_identity["id"]
-        kubelet_identity_principal_id = kubelet_identity["principalId"]
+        kubelet_identity_id = self._get_user_assigned_identity(resource_group, use_for_kubelet=True, paired_control_plane_identity=cluster_identity_id)
+        kubelet_identity_principal_id = self._get_principal_id_of_user_assigned_identity(kubelet_identity_id)
         assert kubelet_identity_id is not None
         assert kubelet_identity_principal_id is not None
         self.kwargs.update(
@@ -14657,7 +14697,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         random_name_length=17,
         name_prefix="clitest",
         location="centraluseuap",
-        preserve_default_location=True,
     )
     def test_aks_and_agentpool_with_static_egress_gateway(
         self, resource_group, resource_group_location
@@ -14772,3 +14811,62 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         self.cmd(delete_cmd, checks=[
             self.is_empty()
         ])
+
+    @AllowLargeResponse()
+    @AKSCustomResourceGroupPreparer(
+        random_name_length=17,
+        name_prefix="clitest",
+        location="centraluseuap",
+    )
+    def test_aks_enable_imds_restriction(
+        self, resource_group, resource_group_location
+    ):
+        _, create_version = self._get_versions(resource_group_location)
+        aks_name = self.create_random_name("cliakstest", 16)
+        self.kwargs.update(
+            {
+                "resource_group": resource_group,
+                "name": aks_name,
+                "location": resource_group_location,
+                "k8s_version": create_version,
+                "ssh_key_value": self.generate_ssh_keys(),
+            }
+        )
+
+        # create
+        create_cmd = (
+            "aks create --resource-group={resource_group} --name={name} --location={location} "
+            "--ssh-key-value={ssh_key_value} "
+            "--kubernetes-version={k8s_version} "
+            "--enable-imds-restriction "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/IMDSRestrictionPreview"
+        )
+        self.cmd(
+            create_cmd,
+            checks=[
+                self.check('provisioningState', 'Succeeded'),
+                self.check(
+                    'networkProfile.podLinkLocalAccess', 'None'
+                ),
+            ],
+        )
+
+        # update -- disable imds restriction
+        update_cmd = (
+            "aks update --resource-group {resource_group} --name {name} "
+            "--disable-imds-restriction --yes "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/IMDSRestrictionPreview "
+            "-o json"
+        )
+        self.cmd(update_cmd, checks=[
+            self.check('provisioningState', 'Succeeded'),
+            self.check(
+                'networkProfile.podLinkLocalAccess', 'IMDS'
+            ),
+        ])
+
+        # delete
+        self.cmd(
+            "aks delete -g {resource_group} -n {name} --yes --no-wait",
+            checks=[self.is_empty()],
+        )