+++ src/aks-preview/HISTORY.rst
@@ -12,9 +12,20 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+7.0.0b2
+++++++++
+* Update the minimum required cli core version to `2.61.0`.
+* Add option `--enable-imds-restriction --disable-imds-restriction` to `az aks create` and `az aks update`. 
+* Introduce valdations to `az aks create` and `az aks update` while using PremiumV2 disk during enabling Azure Container Storage.
+* Delete the Azure Container Storage installation after failure to prevent retries.
+
+7.0.0b1
+++++++++
+* [BREAKING CHANGE]: Remove support for `az aks get-os-options` command.
+
 6.0.0b1
 ++++++++
-* Remove support for `az aks mesh` egress gateway commands.
+* [BREAKING CHANGE]: Remove support for `az aks mesh` egress gateway commands.
 * Add validation to `az aks create` and `az aks update` while modifying the `--ephemeral-disk-volume-type` and `--ephemeral-disk-nvme-perf-tier` values.
 
 5.0.0b4